Get number of occurrences for each taxa of a phyloseq object

<a href="https://adrientaudiere.github.io/MiscMetabar/articles/Rules.html#lifecycle"> <img src="https://img.shields.io/badge/lifecycle-experimental-orange" alt="lifecycle-experimental"></a>

A wrapper of [rgbif::occ_search()] function to get the number of occurences. Optionally, the number of occurrences can be obtained by years or by country.

Usage

tax_gbif_occur_pq(
  physeq = NULL,
  taxnames = NULL,
  taxonomic_rank = "currentCanonicalSimple",
  add_to_phyloseq = NULL,
  col_prefix = NULL,
  by_country = FALSE,
  by_years = FALSE,
  verbose = TRUE,
  time_to_sleep = 0.3,
  discard_genus_alone = identical(taxonomic_rank, "currentCanonicalSimple"),
  discard_NA = TRUE
)

Arguments

physeq

(optional) A phyloseq object. Either `physeq` or `taxnames` must be provided, but not both.

taxnames

(optional) A character vector of taxonomic names.

taxonomic_rank

(Character, default "currentCanonicalSimple") The column(s) present in the @tax_table slot of the phyloseq object. Can be a vector of two columns (e.g. c("Genus", "Species")).

add_to_phyloseq

(logical, default TRUE when physeq is provided, FALSE when taxnames is provided) If TRUE, add new column(s) in the tax_table of the phyloseq object. Automatically set to TRUE when a phyloseq object is provided and FALSE when taxnames is provided. Cannot be TRUE if `taxnames` is provided.

col_prefix

A character string to be added as a prefix to the new columns names added to the tax_table slot of the phyloseq object (default: NULL).

by_country

(logical, default FALSE) If TRUE, the number of occurences is computed by country

by_years

(logical, default FALSE) If TRUE, the number of occurences is computed by years

verbose

(logical, default TRUE) If TRUE, prompt some messages.

time_to_sleep

(numeric, default 0.3) Time to sleep between two calls to rgbif::occ_search(). Useful to avoid to be blocked by GBIF. Try to increase this value if you are blocked by the error "To download GBIF occurrence data in bulk, please request..."

discard_genus_alone

(logical, default `TRUE` when `taxonomic_rank == "currentCanonicalSimple"`). Passed to [taxonomic_rank_to_taxnames()].

discard_NA

(logical, default `TRUE`). Passed to [taxonomic_rank_to_taxnames()].

Value

Either a tibble (if add_to_phyloseq = FALSE) or a new phyloseq object, if add_to_phyloseq = TRUE, with new column(s) in the tax_table.

Details

This function is mainly a wrapper of the work of others. Please cite `rgbif` package.

See also

Author

Adrien Taudiere

Examples

data_fungi_mini_cleanNames <-
  gna_verifier_pq(data_fungi_mini)
#> ✔ GNA verification summary:
#> • Total taxa in phyloseq: 45
#> • Taxa submitted for verification: 37
#> • Genus-level only taxa: 2
#> • Total matches found: 25
#> • Synonyms: 2 (including 2 at genus level)
#> • Accepted names: 23 (including 21 at genus level)

# \donttest{
data_fungi_mini_cleanNames <- tax_gbif_occur_pq(data_fungi_mini_cleanNames, by_country = TRUE)
#> ■■                                 4% | ETA:  7s
#> ℹ Processing GBIF occurrences for Stereum ostrea
#> ■■                                 4% | ETA:  7s
#> ℹ Processing GBIF occurrences for Xylodon raduloides
#> ■■                                 4% | ETA:  7s
#> ■■■■■                             13% | ETA:  8s
#> ℹ Processing GBIF occurrences for Stereum hirsutum
#> ■■■■■                             13% | ETA:  8s
#> ℹ Processing GBIF occurrences for Trametopsis brasiliensis
#> ■■■■■                             13% | ETA:  8s
#> ℹ Processing GBIF occurrences for Basidiodendron eyrei
#> ■■■■■                             13% | ETA:  8s
#> ℹ Processing GBIF occurrences for Sistotrema oblongisporum
#> ■■■■■                             13% | ETA:  8s
#> ℹ Processing GBIF occurrences for Fomes fomentarius
#> ■■■■■                             13% | ETA:  8s
#> ℹ Processing GBIF occurrences for Mycena renati
#> ■■■■■                             13% | ETA:  8s
#> ■■■■■■■■■■■■■                     39% | ETA:  6s
#> ℹ Processing GBIF occurrences for Helicogloea pellucida
#> ■■■■■■■■■■■■■                     39% | ETA:  6s
#> ℹ Processing GBIF occurrences for Radulomyces molaris
#> ■■■■■■■■■■■■■                     39% | ETA:  6s
#> ℹ Processing GBIF occurrences for Elmerina caryae
#> ■■■■■■■■■■■■■                     39% | ETA:  6s
#> ■■■■■■■■■■■■■■■■■                 52% | ETA:  7s
#> ℹ Processing GBIF occurrences for Phanerochaete livescens
#> ■■■■■■■■■■■■■■■■■                 52% | ETA:  7s
#> ℹ Processing GBIF occurrences for Gloeohypochnicium analogum
#> ■■■■■■■■■■■■■■■■■                 52% | ETA:  7s
#> ℹ Processing GBIF occurrences for Hyphoderma roseocremeum
#> ■■■■■■■■■■■■■■■■■                 52% | ETA:  7s
#> ℹ Processing GBIF occurrences for Hyphoderma setigerum
#> ■■■■■■■■■■■■■■■■■                 52% | ETA:  7s
#> ℹ Processing GBIF occurrences for Trametes versicolor
#> ■■■■■■■■■■■■■■■■■                 52% | ETA:  7s
#> ℹ Processing GBIF occurrences for Peniophora versiformis
#> ■■■■■■■■■■■■■■■■■                 52% | ETA:  7s
#> ■■■■■■■■■■■■■■■■■■■■■■■■          78% | ETA:  3s
#> ℹ Processing GBIF occurrences for Exidia glandulosa
#> ■■■■■■■■■■■■■■■■■■■■■■■■          78% | ETA:  3s
#> ℹ Processing GBIF occurrences for Peniophorella pubera
#> ■■■■■■■■■■■■■■■■■■■■■■■■          78% | ETA:  3s
#> ℹ Processing GBIF occurrences for Auricularia mesenterica
#> ■■■■■■■■■■■■■■■■■■■■■■■■          78% | ETA:  3s
#> ℹ Processing GBIF occurrences for Laetisaria buckii
#> ■■■■■■■■■■■■■■■■■■■■■■■■          78% | ETA:  3s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    96% | ETA:  1s
#> ℹ Processing GBIF occurrences for Hericium coralloides
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    96% | ETA:  1s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> 
#> ℹ Processing GBIF occurrences for Xylodon flaviporus

# Get data without adding to phyloseq
tax_gbif_occur_pq(data_fungi_mini_cleanNames, add_to_phyloseq = FALSE)
#> ℹ Processing GBIF occurrences for Stereum ostrea
#> ■■■■                               9% | ETA: 18s
#> ℹ Processing GBIF occurrences for Xylodon raduloides
#> ■■■■                               9% | ETA: 18s
#> ℹ Processing GBIF occurrences for Stereum hirsutum
#> ■■■■                               9% | ETA: 18s
#> ℹ Processing GBIF occurrences for Trametopsis brasiliensis
#> ■■■■                               9% | ETA: 18s
#> ℹ Processing GBIF occurrences for Basidiodendron eyrei
#> ■■■■                               9% | ETA: 18s
#> ■■■■■■■■■                         26% | ETA: 13s
#> ℹ Processing GBIF occurrences for Sistotrema oblongisporum
#> ■■■■■■■■■                         26% | ETA: 13s
#> ℹ Processing GBIF occurrences for Fomes fomentarius
#> ■■■■■■■■■                         26% | ETA: 13s
#> ℹ Processing GBIF occurrences for Mycena renati
#> ■■■■■■■■■                         26% | ETA: 13s
#> ℹ Processing GBIF occurrences for Helicogloea pellucida
#> ■■■■■■■■■                         26% | ETA: 13s
#> ■■■■■■■■■■■■■■                    43% | ETA: 10s
#> ℹ Processing GBIF occurrences for Radulomyces molaris
#> ■■■■■■■■■■■■■■                    43% | ETA: 10s
#> ℹ Processing GBIF occurrences for Elmerina caryae
#> ■■■■■■■■■■■■■■                    43% | ETA: 10s
#> ℹ Processing GBIF occurrences for Phanerochaete livescens
#> ■■■■■■■■■■■■■■                    43% | ETA: 10s
#> ℹ Processing GBIF occurrences for Gloeohypochnicium analogum
#> ■■■■■■■■■■■■■■                    43% | ETA: 10s
#> ■■■■■■■■■■■■■■■■■■■               61% | ETA:  7s
#> ℹ Processing GBIF occurrences for Hyphoderma roseocremeum
#> ■■■■■■■■■■■■■■■■■■■               61% | ETA:  7s
#> ℹ Processing GBIF occurrences for Hyphoderma setigerum
#> ■■■■■■■■■■■■■■■■■■■               61% | ETA:  7s
#> ℹ Processing GBIF occurrences for Trametes versicolor
#> ■■■■■■■■■■■■■■■■■■■               61% | ETA:  7s
#> ■■■■■■■■■■■■■■■■■■■■■■■           74% | ETA:  5s
#> ℹ Processing GBIF occurrences for Peniophora versiformis
#> ■■■■■■■■■■■■■■■■■■■■■■■           74% | ETA:  5s
#> ℹ Processing GBIF occurrences for Exidia glandulosa
#> ■■■■■■■■■■■■■■■■■■■■■■■           74% | ETA:  5s
#> ℹ Processing GBIF occurrences for Peniophorella pubera
#> ■■■■■■■■■■■■■■■■■■■■■■■           74% | ETA:  5s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       87% | ETA:  3s
#> ℹ Processing GBIF occurrences for Auricularia mesenterica
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       87% | ETA:  3s
#> ℹ Processing GBIF occurrences for Laetisaria buckii
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       87% | ETA:  3s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    96% | ETA:  1s
#> ℹ Processing GBIF occurrences for Hericium coralloides
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■    96% | ETA:  1s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> 
#> ℹ Processing GBIF occurrences for Xylodon flaviporus
#> # A tibble: 23 × 2
#>    Global_occurences canonicalName           
#>                <int> <chr>                   
#>  1             11508 Stereum ostrea          
#>  2              5651 Xylodon raduloides      
#>  3            122976 Stereum hirsutum        
#>  4                 8 Trametopsis brasiliensis
#>  5              1165 Basidiodendron eyrei    
#>  6              3050 Sistotrema oblongisporum
#>  7            172793 Fomes fomentarius       
#>  8              5785 Mycena renati           
#>  9                43 Helicogloea pellucida   
#> 10              4714 Radulomyces molaris     
#> # ℹ 13 more rows
tax_gbif_occur_pq(data_fungi_mini_cleanNames, by_years = TRUE, add_to_phyloseq = FALSE)
#> ℹ Processing GBIF occurrences for Stereum ostrea
#> ℹ Processing GBIF occurrences for Xylodon raduloides
#> ■■■■■                             13% | ETA:  9s
#> ℹ Processing GBIF occurrences for Stereum hirsutum
#> ■■■■■                             13% | ETA:  9s
#> ℹ Processing GBIF occurrences for Trametopsis brasiliensis
#> ■■■■■                             13% | ETA:  9s
#> ℹ Processing GBIF occurrences for Basidiodendron eyrei
#> ■■■■■                             13% | ETA:  9s
#> ℹ Processing GBIF occurrences for Sistotrema oblongisporum
#> ■■■■■                             13% | ETA:  9s
#> ℹ Processing GBIF occurrences for Fomes fomentarius
#> ■■■■■                             13% | ETA:  9s
#> ■■■■■■■■■■■                       35% | ETA:  8s
#> ℹ Processing GBIF occurrences for Mycena renati
#> ■■■■■■■■■■■                       35% | ETA:  8s
#> ℹ Processing GBIF occurrences for Helicogloea pellucida
#> ■■■■■■■■■■■                       35% | ETA:  8s
#> ℹ Processing GBIF occurrences for Radulomyces molaris
#> ■■■■■■■■■■■                       35% | ETA:  8s
#> ℹ Processing GBIF occurrences for Elmerina caryae
#> ■■■■■■■■■■■                       35% | ETA:  8s
#> ℹ Processing GBIF occurrences for Phanerochaete livescens
#> ■■■■■■■■■■■                       35% | ETA:  8s
#> ■■■■■■■■■■■■■■■■■■                57% | ETA:  5s
#> ℹ Processing GBIF occurrences for Gloeohypochnicium analogum
#> ■■■■■■■■■■■■■■■■■■                57% | ETA:  5s
#> ℹ Processing GBIF occurrences for Hyphoderma roseocremeum
#> ■■■■■■■■■■■■■■■■■■                57% | ETA:  5s
#> ℹ Processing GBIF occurrences for Hyphoderma setigerum
#> ■■■■■■■■■■■■■■■■■■                57% | ETA:  5s
#> ℹ Processing GBIF occurrences for Trametes versicolor
#> ■■■■■■■■■■■■■■■■■■                57% | ETA:  5s
#> ℹ Processing GBIF occurrences for Peniophora versiformis
#> ■■■■■■■■■■■■■■■■■■                57% | ETA:  5s
#> ℹ Processing GBIF occurrences for Exidia glandulosa
#> ■■■■■■■■■■■■■■■■■■                57% | ETA:  5s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■        83% | ETA:  2s
#> ℹ Processing GBIF occurrences for Peniophorella pubera
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■        83% | ETA:  2s
#> ℹ Processing GBIF occurrences for Auricularia mesenterica
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■        83% | ETA:  2s
#> ℹ Processing GBIF occurrences for Laetisaria buckii
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■        83% | ETA:  2s
#> ℹ Processing GBIF occurrences for Hericium coralloides
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■        83% | ETA:  2s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> 
#> ℹ Processing GBIF occurrences for Xylodon flaviporus
#> # A tibble: 23 × 57
#> # Groups:   canonicalName [23]
#>    canonicalName  `2024` `2025` `2023` `2022` `2026` `2021` `2012` `2013` `2018`
#>    <chr>           <int>  <int>  <int>  <int>  <int>  <int>  <int>  <int>  <int>
#>  1 Stereum ostrea    932    883    724    361    302    213    188    175    173
#>  2 Xylodon radul…     NA     NA     NA     NA     NA     NA    158     NA    179
#>  3 Stereum hirsu…   6263   6655   7077   5318     NA   5093   3147     NA   3048
#>  4 Trametopsis b…     NA     NA     NA     NA     NA     NA     NA     NA     NA
#>  5 Basidiodendro…     NA     NA     NA     NA     NA     NA     NA     30     42
#>  6 Sistotrema ob…     NA     NA     NA     NA     NA    152     83    302     74
#>  7 Fomes fomenta…  19330  22956  18705  17585   6683  13194     NA     NA   4470
#>  8 Mycena renati     454    428    387    308     NA    275     NA    154    231
#>  9 Helicogloea p…     NA     NA      2     NA     NA     NA     NA     NA      1
#> 10 Radulomyces m…     NA    141    139     NA     NA     NA     NA     NA    145
#> # ℹ 13 more rows
#> # ℹ 47 more variables: `2005` <int>, `2019` <int>, `2006` <int>, `2007` <int>,
#> #   `2016` <int>, `2020` <int>, `2017` <int>, `2000` <int>, `1992` <int>,
#> #   `1999` <int>, `1979` <int>, `1990` <int>, `1997` <int>, `1939` <int>,
#> #   `1936` <int>, `1998` <int>, `1956` <int>, `2014` <int>, `2015` <int>,
#> #   `2009` <int>, `1941` <int>, `2010` <int>, `2008` <int>, `2002` <int>,
#> #   `2004` <int>, `1901` <int>, `1975` <int>, `1996` <int>, `2001` <int>, …

# Using taxnames vector (returns a tibble)
tax_gbif_occur_pq(taxnames = c("Amanita muscaria", "Boletus edulis"))
#> ℹ Processing GBIF occurrences for Amanita muscaria
#> ℹ Processing GBIF occurrences for Boletus edulis
#> # A tibble: 2 × 2
#>   Global_occurences canonicalName   
#>               <int> <chr>           
#> 1            316020 Amanita muscaria
#> 2             79350 Boletus edulis  
ggplot(
  data_fungi_mini_cleanNames@tax_table,
  aes(y = log10(as.numeric(Global_occurences)), x = currentCanonicalSimple)
) +
  geom_col() +
  geom_col(aes(y = -log10(as.numeric(FR))), fill = "blue") +
  coord_flip() +
  xlab("Number of occurences (log10 scale) at global (grey) scale and in France (blue)")
#> Error in geom_col(): Problem while computing aesthetics.
#> ℹ Error occurred in the 1st layer.
#> Caused by error:
#> ! object 'Global_occurences' not found
# }