Find photos of taxa from GBIF or Wikitaxa

Find photos of taxa from GBIF or Wikitaxa

Usage

tax_photos_pq(
  physeq = NULL,
  taxnames = NULL,
  taxonomic_rank = "currentCanonicalSimple",
  source = "gbif",
  folder_name = "photos_physeq",
  add_to_phyloseq = NULL,
  col_prefix = NULL,
  gallery = FALSE,
  overwrite_folder = FALSE,
  col_name_url = "photo_url",
  verbose = TRUE,
  caption_valign = "bottom",
  caption_font_size = 12,
  simple_caption = FALSE,
  ...
)

Arguments

physeq

(optional) A phyloseq object. Either `physeq` or `taxnames` must be provided, but not both.

taxnames

(optional) A character vector of taxonomic names.

taxonomic_rank

(Character, default = "currentCanonicalSimple") The column(s) present in the @tax_table slot of the phyloseq object. Can be a vector of two columns (e.g. the c("Genus", "Species")).

source

(Character) either "gbif" or "wikitaxa".

folder_name

(default "photos_physeq") Name of the folder where photos will be downloaded. Only used if both add_to_phyloseq and gallery are FALSE.

add_to_phyloseq

(logical, default TRUE when physeq is provided, FALSE when taxnames is provided) If TRUE, a new phyloseq object is returned with a new column containing the URL (entitled with the parameter col_name_url) in the tax_table. Automatically set to TRUE when a phyloseq object is provided and FALSE when taxnames is provided. Cannot be TRUE if `taxnames` is provided.

col_prefix

A character string to be added as a prefix to the new columns names added to the tax_table slot of the phyloseq object (default: NULL).

gallery

(logical, default FALSE) If TRUE, a html gallery is created using the function [pixture::pixgallery()].

overwrite_folder

(logical, default FALSE) If TRUE, the folder specified in the parameter folder_name will be deleted if it already exists.

col_name_url

(default "photo_url") Name of the new column in the tax_table

verbose

(logical, default TRUE) If TRUE, prompt some messages.

caption_valign

(character, default "bottom") Vertical alignment of the caption in the gallery.

caption_font_size

(int) Size of the caption font in the gallery.

simple_caption

(logical, default FALSE) If TRUE, the caption of the gallery photo will be only the taxonomic name. If FALSE, the caption include information from the phyloseq object (number of sequences, taxa and samples).

...

Other parameters to be passed to pixture::pixgallery() function.

Value

There is three behavior.(i) If gallery = TRUE, a html gallery is created using the function [pixture::pixgallery()]. (ii) If add_to_phyloseq = TRUE, a new phyloseq object is returned with a new column (called with the parameter col_name_url) in the tax_table containing the URL. (iii) If both gallery = FALSE and add_to_phyloseq = FALSE, photos are downloaded in a folder (folder_name parameter) and the list of url are returned in the form of a tibble.

Details

There is three behavior. See the returns section. Gbif source is quicker than wikitaxa source. Note that for the moment the function only return one photo per species.

Author

Adrien Taudiere

Examples

data_fungi_mini_cleanNames <- gna_verifier_pq(data_fungi_mini)
#> ✔ GNA verification summary:
#> • Total taxa in phyloseq: 45
#> • Taxa submitted for verification: 37
#> • Genus-level only taxa: 2
#> • Total matches found: 25
#> • Synonyms: 2 (including 2 at genus level)
#> • Accepted names: 23 (including 21 at genus level)

tax_photos_pq(data_fungi_mini_cleanNames,
  gallery = TRUE,
  h = "40px",
  w = "80px",
  source = "wikitaxa"
)
#> ■■                                 4% | ETA:  0s
#> ℹ 1/23 - Downloading photo of Stereum ostrea
#> ■■                                 4% | ETA:  0s
#> ℹ 2/23 - No photo available for Xylodon raduloides
#> ■■                                 4% | ETA:  0s
#> ■■■■■                             13% | ETA: 21s
#> ℹ 3/23 - Downloading photo of Stereum hirsutum
#> ■■■■■                             13% | ETA: 21s
#> ℹ 4/23 - No photo available for Trametopsis brasiliensis
#> ■■■■■                             13% | ETA: 21s
#> ■■■■■■■■                          22% | ETA: 22s
#> ℹ 5/23 - No photo available for Basidiodendron eyrei
#> ■■■■■■■■                          22% | ETA: 22s
#> ℹ 6/23 - No photo available for Sistotrema oblongisporum
#> ■■■■■■■■                          22% | ETA: 22s
#> ℹ 7/23 - Downloading photo of Fomes fomentarius
#> ■■■■■■■■                          22% | ETA: 22s
#> ■■■■■■■■■■■                       35% | ETA: 19s
#> ℹ 8/23 - Downloading photo of Mycena renati
#> ■■■■■■■■■■■                       35% | ETA: 19s
#> ■■■■■■■■■■■■■                     39% | ETA: 18s
#> ℹ 9/23 - No photo available for Helicogloea pellucida
#> ■■■■■■■■■■■■■                     39% | ETA: 18s
#> ℹ 10/23 - No photo available for Radulomyces molaris
#> ■■■■■■■■■■■■■                     39% | ETA: 18s
#> ℹ 11/23 - No photo available for Elmerina caryae
#> ■■■■■■■■■■■■■                     39% | ETA: 18s
#> ■■■■■■■■■■■■■■■■■                 52% | ETA: 14s
#> ℹ 12/23 - No photo available for Phanerochaete livescens
#> ■■■■■■■■■■■■■■■■■                 52% | ETA: 14s
#> ℹ 13/23 - No photo available for Gloeohypochnicium analogum
#> ■■■■■■■■■■■■■■■■■                 52% | ETA: 14s
#> ℹ 14/23 - No photo available for Hyphoderma roseocremeum
#> ■■■■■■■■■■■■■■■■■                 52% | ETA: 14s
#> ■■■■■■■■■■■■■■■■■■■■■             65% | ETA: 10s
#> ℹ 15/23 - No photo available for Hyphoderma setigerum
#> ■■■■■■■■■■■■■■■■■■■■■             65% | ETA: 10s
#> ℹ 16/23 - Downloading photo of Trametes versicolor
#> ■■■■■■■■■■■■■■■■■■■■■             65% | ETA: 10s
#> ■■■■■■■■■■■■■■■■■■■■■■■           74% | ETA:  8s
#> ℹ 17/23 - No photo available for Peniophora versiformis
#> ■■■■■■■■■■■■■■■■■■■■■■■           74% | ETA:  8s
#> ■■■■■■■■■■■■■■■■■■■■■■■■          78% | ETA:  7s
#> ℹ 18/23 - Downloading photo of Exidia glandulosa
#> ■■■■■■■■■■■■■■■■■■■■■■■■          78% | ETA:  7s
#> ℹ 19/23 - No photo available for Peniophorella pubera
#> ■■■■■■■■■■■■■■■■■■■■■■■■          78% | ETA:  7s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       87% | ETA:  4s
#> ℹ 20/23 - No photo available for Auricularia mesenterica
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■       87% | ETA:  4s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      91% | ETA:  3s
#> ℹ 21/23 - No photo available for Laetisaria buckii
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      91% | ETA:  3s
#> ℹ 22/23 - Downloading photo of Hericium coralloides
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■      91% | ETA:  3s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> 
#> ℹ 23/23 - No photo available for Xylodon flaviporus
#> ✔ Photo download summary:/n
#> - 7 photos found and downloaded/n
#> - 13 taxa depicted/n
#> - 16 taxonomic names not found/n
#> - 32 taxa have no photo URL
#> ℹ Creating captions for gallery

tax_photos_pq(
  taxnames = c("Xylodon flaviporus", "Basidiodendron eyrei"),
  gallery = TRUE,
  layout = "rhombus"
)
#> ■■■■■■■■■■■■■■■■                  50% | ETA:  0s
#> ℹ 1/2 - Downloading photo of Xylodon flaviporus
#> ■■■■■■■■■■■■■■■■                  50% | ETA:  0s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> 
#> ℹ 2/2 - Downloading photo of Basidiodendron eyrei
#> ✔ Photo download summary:/n
#> - 2 photos found and downloaded/n
#> - 0 taxonomic names not found
#> ℹ Creating captions for gallery

data_fungi_mini_cleanNames_photos <-
  tax_photos_pq(data_fungi_mini_cleanNames)
#> ℹ 1/23 - No photo available for Stereum ostrea
#> ■■■■                               9% | ETA:  1s
#> ℹ 2/23 - No photo available for Xylodon raduloides
#> ■■■■                               9% | ETA:  1s
#> ℹ 3/23 - Downloading photo of Stereum hirsutum
#> ■■■■                               9% | ETA:  1s
#> ℹ 4/23 - No photo available for Trametopsis brasiliensis
#> ■■■■                               9% | ETA:  1s
#> ℹ 5/23 - Downloading photo of Basidiodendron eyrei
#> ■■■■                               9% | ETA:  1s
#> ℹ 6/23 - No photo available for Sistotrema oblongisporum
#> ■■■■                               9% | ETA:  1s
#> ℹ 7/23 - Downloading photo of Fomes fomentarius
#> ■■■■                               9% | ETA:  1s
#> ℹ 8/23 - Downloading photo of Mycena renati
#> ■■■■                               9% | ETA:  1s
#> ℹ 9/23 - No photo available for Helicogloea pellucida
#> ■■■■                               9% | ETA:  1s
#> ℹ 10/23 - Downloading photo of Radulomyces molaris
#> ■■■■                               9% | ETA:  1s
#> ℹ 11/23 - No photo available for Elmerina caryae
#> ■■■■                               9% | ETA:  1s
#> ℹ 12/23 - No photo available for Phanerochaete livescens
#> ■■■■                               9% | ETA:  1s
#> ℹ 13/23 - Downloading photo of Gloeohypochnicium analogum
#> ■■■■                               9% | ETA:  1s
#> ℹ 14/23 - Downloading photo of Hyphoderma roseocremeum
#> ■■■■                               9% | ETA:  1s
#> ℹ 15/23 - Downloading photo of Hyphoderma setigerum
#> ■■■■                               9% | ETA:  1s
#> ℹ 16/23 - Downloading photo of Trametes versicolor
#> ■■■■                               9% | ETA:  1s
#> ℹ 17/23 - No photo available for Peniophora versiformis
#> ■■■■                               9% | ETA:  1s
#> ℹ 18/23 - Downloading photo of Exidia glandulosa
#> ■■■■                               9% | ETA:  1s
#> ℹ 19/23 - Downloading photo of Peniophorella pubera
#> ■■■■                               9% | ETA:  1s
#> ℹ 20/23 - Downloading photo of Auricularia mesenterica
#> ■■■■                               9% | ETA:  1s
#> ℹ 21/23 - No photo available for Laetisaria buckii
#> ■■■■                               9% | ETA:  1s
#> ℹ 22/23 - Downloading photo of Hericium coralloides
#> ■■■■                               9% | ETA:  1s
#> ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> 
#> ℹ 23/23 - Downloading photo of Xylodon flaviporus
#> ✔ Photo download summary:/n
#> - 14 photos found and downloaded/n
#> - 18 taxa depicted/n
#> - 9 taxonomic names not found/n
#> - 27 taxa have no photo URL

# Which photo(s) depicted more than one OTU
data_fungi_mini_cleanNames_photos@tax_table[, "photo_url"] |>
  table() |>
  (\(tab) tab[as.numeric(tab) > 1])()
#> 
#> http://www.mycokey.com/MycoKeySolidState/pictures/basi/hete/Exidia/glan10L.jpg 
#>                                                                              2 
#>                                https://images.naturalis.nl/original/191145.jpg 
#>                                                                              4